Fix using system parameters in reactions
Previously, accessing the system parameters (like porosity) in the
reactions lead to errors because they can be an array with a spatial
axis, but this was not mapped correctly with vmap when evaluating the
reaction right-hand-side.

This commit also enables the user to add custom system parameters that
can be accessed in all reactions (e.g., solid density).